Fix realloc bug
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@24179 9d14d920-75d5-0310-8bd7-a52572981aee

+++ agents/unix/radius.c
@@ -594,7 +594,7 @@ expand_rp(const char *value, radius_parameter *top)
         {
             const char *rp_val;
             size_t      rpv_len;
-            size_t      diff_len;
+            int         diff_len;
 
             *next = '\0';
             if (!retrieve_rp(top, new_ptr + 2, &rp_val))
